A natural language processing project with using Python
The aim of the project is to contribute bioinformatics to accelerating vaccine developments. The human immune system develops a defense against foreign substances (pathogens) that enter the body. Developed vaccines greatly help this system to develop defenses against pathogens. While each pathogen has its own antigen region, the body also produces specific antibodies that fight these regions. For this reason, spike cells, which are copies of the pathogen, help the body produce antibodies if found in vaccines. In the project, it is aimed to predict the specific antigen region by applying natural language processing methods to the sequence of amino acids represented by 20 different letters of the relevant pathogen.